refactor: community.block 중복 임시 컴포넌트 제거
