+++ app-main/src/main/java/net/causw/app/main/domain/community/block/api/v2/mapper/ChildCommentBlockDtoMapper.java
@@ -3,8 +3,8 @@ package net.causw.app.main.domain.community.block.api.v2.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
-import net.causw.app.main.domain.community.block.api.v2.dto.response.BlockResponse;
-import net.causw.app.main.domain.community.block.service.dto.BlockCreateResult;
+import net.causw.app.main.domain.user.account.api.v2.dto.response.BlockResponseDto;
+import net.causw.app.main.domain.user.relation.service.v2.dto.BlockCreateResult;
 import net.causw.app.main.domain.community.block.service.dto.ChildCommentBlockCreateCommand;
 import net.causw.app.main.domain.user.account.entity.user.User;
 
@@ -14,5 +14,5 @@ public interface ChildCommentBlockDtoMapper {
 	@Mapping(target = "blocker", source = "user")
 	ChildCommentBlockCreateCommand toCommand(String targetUserId, String childCommentId, User user);
 
-	BlockResponse toResponse(BlockCreateResult result);
+	BlockResponseDto toResponse(BlockCreateResult result);
 }
